Removes overriding the visibility of ActionController::Base#cookies
Rails declares ActionController::Base#cookies to be a private method.
CASino should respect this and not change it to be public just so that
listeners can interact with them.

This adds a convenience method to CASino::Listener that exposes an API
into ActionController::Base#cookies by calling #send in order to access
the private method.